+++ src/DenseMatrix.cpp
@@ -18,7 +18,7 @@
 namespace seamat {
 // Parameter Constructor
 template<typename T>
-DenseMatrix<T>::DenseMatrix(uint32_t _rows, uint32_t _cols, const T& _initial) {
+DenseMatrix<T>::DenseMatrix(size_t _rows, size_t _cols, const T& _initial) {
     mat.resize(_rows*_cols, _initial);
     this->resize_rows(_rows);
     this->resize_cols(_cols);
@@ -26,7 +26,7 @@ DenseMatrix<T>::DenseMatrix(uint32_t _rows, uint32_t _cols, const T& _initial) {
 
 // Copy constructor from contiguous 2D vector
 template <typename T>
-DenseMatrix<T>::DenseMatrix(const std::vector<T> &rhs, const uint32_t _rows, const uint32_t _cols) {
+DenseMatrix<T>::DenseMatrix(const std::vector<T> &rhs, const size_t _rows, const size_t _cols) {
     mat = rhs;
     this->resize_rows(_rows);
     this->resize_cols(_cols);
@@ -39,8 +39,8 @@ DenseMatrix<T>::DenseMatrix(const std::vector<std::vector<T>> &rhs) {
     this->resize_cols(rhs.at(0).size());
     mat.resize(this->get_rows()*this->get_cols());
 #pragma omp parallel for schedule(static)
-    for (uint32_t i = 0; i < this->get_rows(); ++i) {
-	for (uint32_t j = 0; j < this->get_cols(); ++j) {
+    for (size_t i = 0; i < this->get_rows(); ++i) {
+	for (size_t j = 0; j < this->get_cols(); ++j) {
 	    this->operator()(i, j) = rhs[i][j];
 	}
     }
@@ -53,8 +53,8 @@ DenseMatrix<T>::DenseMatrix(const Matrix<T> &rhs) {
     this->resize_cols(rhs.get_cols());
     mat.resize(this->get_rows()*this->get_cols());
 #pragma omp parallel for schedule(static)
-    for (uint32_t i = 0; i < this->get_rows(); ++i) {
-	for (uint32_t j = 0; j < this->get_cols(); ++j) {
+    for (size_t i = 0; i < this->get_rows(); ++i) {
+	for (size_t j = 0; j < this->get_cols(); ++j) {
 	    this->operator()(i, j) = rhs(i, j);
 	}
     }
@@ -66,14 +66,14 @@ DenseMatrix<T>& DenseMatrix<T>::operator=(const Matrix<T>& rhs) {
     if (&rhs == this)
 	return *this;
 
-    uint32_t new_rows = rhs.get_rows();
-    uint32_t new_cols = rhs.get_cols();
+    size_t new_rows = rhs.get_rows();
+    size_t new_cols = rhs.get_cols();
     if (new_rows != this->get_rows() || new_cols != this->get_cols()) {
 	resize(new_rows, new_cols, (T)0);
     }
 #pragma omp parallel for schedule(static)
-    for (uint32_t i = 0; i < new_rows; i++) {
-	for (uint32_t j = 0; j < new_cols; j++) {
+    for (size_t i = 0; i < new_rows; i++) {
+	for (size_t j = 0; j < new_cols; j++) {
 	    this->operator()(i, j) = rhs(i, j);
 	}
     }
@@ -82,7 +82,7 @@ DenseMatrix<T>& DenseMatrix<T>::operator=(const Matrix<T>& rhs) {
 
 // Resize a matrix
 template<typename T>
-void DenseMatrix<T>::resize(const uint32_t new_rows, const uint32_t new_cols, const T initial) {
+void DenseMatrix<T>::resize(const size_t new_rows, const size_t new_cols, const T initial) {
     if (new_rows != this->get_rows() || new_cols != this->get_cols()) {
 	mat.resize(new_rows*new_cols, initial);
 	this->resize_rows(new_rows);
@@ -92,13 +92,13 @@ void DenseMatrix<T>::resize(const uint32_t new_rows, const uint32_t new_cols, co
 
 // Access individual elements
 template <typename T>
-T& DenseMatrix<T>::operator()(uint32_t row, uint32_t col) {
+T& DenseMatrix<T>::operator()(size_t row, size_t col) {
     return this->mat[row*this->get_cols() + col];
 }
 
 // Access individual elements (const)
 template <typename T>
-const T& DenseMatrix<T>::operator()(uint32_t row, uint32_t col) const {
+const T& DenseMatrix<T>::operator()(size_t row, size_t col) const {
     return this->mat[row*this->get_cols() + col];
 }
